🚚(frontend) rename pad to pad-editor
Rename the pad feature to pad-editor.
