+++ src/frontend/apps/impress/src/features/pads/pad-editor/components/PadEditor.tsx
