+++ src/frontend/apps/impress/src/features/pads/pad-editor/stores/index.ts
