+++ src/frontend/apps/impress/src/features/pads/pad-editor/utils.ts
